requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic-settings==2.2.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
boto3==1.34.81
redis==5.0.3
rq==1.15.1
python-magic==0.4.27
PyJWT==2.8.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
black==24.3.0
ruff==0.3.4

